% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoders.R
\name{decoder_byte_level}
\alias{decoder_byte_level}
\title{Byte level decoder}
\description{
Byte level decoder

Byte level decoder
}
\details{
This decoder is to be used with the \link{pre_tokenizer_byte_level}.
}
\seealso{
Other decoders: 
\code{\link{tok_decoder}}
}
\concept{decoders}
\section{Super class}{
\code{\link[tok:tok_decoder]{tok::tok_decoder}} -> \code{tok_decoder_byte_level}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tok_decoder_byte_level-new}{\code{decoder_byte_level$new()}}
\item \href{#method-tok_decoder_byte_level-clone}{\code{decoder_byte_level$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_decoder_byte_level-new"></a>}}
\if{latex}{\out{\hypertarget{method-tok_decoder_byte_level-new}{}}}
\subsection{Method \code{new()}}{
Initializes a byte level decoder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decoder_byte_level$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_decoder_byte_level-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tok_decoder_byte_level-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decoder_byte_level$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
