% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizers.R
\name{normalizer_nfkc}
\alias{normalizer_nfkc}
\title{NFKC normalizer}
\description{
NFKC normalizer

NFKC normalizer
}
\seealso{
Other normalizers: 
\code{\link{normalizer_nfc}},
\code{\link{tok_normalizer}}
}
\concept{normalizers}
\section{Super class}{
\code{\link[tok:tok_normalizer]{tok::tok_normalizer}} -> \code{tok_normalizer_nfc}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tok_normalizer_nfc-new}{\code{normalizer_nfkc$new()}}
\item \href{#method-tok_normalizer_nfc-clone}{\code{normalizer_nfkc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_normalizer_nfc-new"></a>}}
\if{latex}{\out{\hypertarget{method-tok_normalizer_nfc-new}{}}}
\subsection{Method \code{new()}}{
Initializes the NFKC normalizer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{normalizer_nfkc$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_normalizer_nfc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tok_normalizer_nfc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{normalizer_nfkc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
