% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_tokenizers.R
\name{pre_tokenizer}
\alias{pre_tokenizer}
\title{Generic class for tokenizers}
\description{
Generic class for tokenizers

Generic class for tokenizers
}
\seealso{
Other pre_tokenizer: 
\code{\link{pre_tokenizer_byte_level}},
\code{\link{pre_tokenizer_whitespace}}
}
\concept{pre_tokenizer}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.pre_tokenizer}}{Internal pointer to tokenizer object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tok_pre_tokenizer-new}{\code{pre_tokenizer$new()}}
\item \href{#method-tok_pre_tokenizer-clone}{\code{pre_tokenizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_pre_tokenizer-new"></a>}}
\if{latex}{\out{\hypertarget{method-tok_pre_tokenizer-new}{}}}
\subsection{Method \code{new()}}{
Initializes a tokenizer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pre_tokenizer$new(pre_tokenizer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pre_tokenizer}}{a raw pointer to a tokenizer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_pre_tokenizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tok_pre_tokenizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pre_tokenizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
