% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{tok-package}
\alias{tok}
\alias{tok-package}
\title{tok: Fast Text Tokenization}
\description{
Interfaces with the 'Hugging Face' tokenizers library to provide implementations of today's most used tokenizers such as the 'Byte-Pair Encoding' algorithm \url{https://huggingface.co/docs/tokenizers/index}. It's extremely fast for both training new vocabularies and tokenizing texts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlverse/tok}
  \item Report bugs at \url{https://github.com/mlverse/tok/issues}
}

}
\author{
\strong{Maintainer}: Daniel Falbel \email{daniel@posit.co}

Other contributors:
\itemize{
  \item  Posit [copyright holder]
}

}
