% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoders.R
\name{tok_decoder}
\alias{tok_decoder}
\title{Generic class for decoders}
\description{
Generic class for decoders

Generic class for decoders
}
\seealso{
Other decoders: 
\code{\link{decoder_byte_level}}
}
\concept{decoders}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.decoder}}{The raw pointer to the decoder}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tok_decoder-new}{\code{tok_decoder$new()}}
\item \href{#method-tok_decoder-clone}{\code{tok_decoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_decoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-tok_decoder-new}{}}}
\subsection{Method \code{new()}}{
Initializes a decoder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tok_decoder$new(decoder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{decoder}}{a raw decoder pointer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_decoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tok_decoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tok_decoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
