% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processors.R
\name{tok_processor}
\alias{tok_processor}
\title{Generic class for processors}
\description{
Generic class for processors

Generic class for processors
}
\seealso{
Other processors: 
\code{\link{processor_byte_level}}
}
\concept{processors}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.processor}}{Internal pointer to processor object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tok_processor-new}{\code{tok_processor$new()}}
\item \href{#method-tok_processor-clone}{\code{tok_processor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_processor-new"></a>}}
\if{latex}{\out{\hypertarget{method-tok_processor-new}{}}}
\subsection{Method \code{new()}}{
Initializes a tokenizer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tok_processor$new(processor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{processor}}{a raw pointer to a processor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_processor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tok_processor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tok_processor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
