% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainers.R
\name{tok_trainer}
\alias{tok_trainer}
\title{Generic training class}
\description{
Generic training class

Generic training class
}
\seealso{
Other trainer: 
\code{\link{trainer_bpe}},
\code{\link{trainer_unigram}},
\code{\link{trainer_wordpiece}}
}
\concept{trainer}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.trainer}}{a pointer to a raw trainer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tok_trainer-new}{\code{tok_trainer$new()}}
\item \href{#method-tok_trainer-clone}{\code{tok_trainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_trainer-new"></a>}}
\if{latex}{\out{\hypertarget{method-tok_trainer-new}{}}}
\subsection{Method \code{new()}}{
Initializes a generic trainer from a raw trainer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tok_trainer$new(trainer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trainer}}{raw trainer (internal)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tok_trainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tok_trainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tok_trainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
